requirements.txt
Flask==3.0.0
flask-cors==6.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
openai==0.28.1